% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatialNiche.R
\name{BuildNicheExpressionAssay}
\alias{BuildNicheExpressionAssay}
\title{Build Niche Expression Assay}
\usage{
BuildNicheExpressionAssay(data, coords, neighbors.k = 30)
}
\arguments{
\item{data}{gene x cell matrix containing expression values}

\item{coords}{cell x 2 matrix containing spatial coordinate of cells}

\item{neighbors.k}{nearest neighbours to use}
}
\value{
matrix of nearest neighbour aggregated counts for each feature, excluding the cell itself
}
\description{
Creates a matrix of expression values for each cell's nearest spatial neighbours
To be used as an alternative niche assay without needing to specify cell types apriori - can test for enrichment afterwards
So this is a more unsupervised approach
}
