% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatialNiche.R
\name{CellTypeNicheEnrichment}
\alias{CellTypeNicheEnrichment}
\title{Cell Type Niche Enrichment}
\usage{
CellTypeNicheEnrichment(niches, cell.types)
}
\arguments{
\item{niches}{- vector of niche names per cell}

\item{cell.types}{- vector of cell types per cell, in the same cell order as that of niches}
}
\value{
data frame with enrichment estimate and p value for each  cell type - cell niche combination
}
\description{
Given a vector of niches and a corresponding vector of cell types, will calculate cell type enrichment table in each niche
to be used with downstream visualisation functions
}
