% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualiseProjection.R
\name{CompareClusters}
\alias{CompareClusters}
\title{Compare Clusters}
\usage{
CompareClusters(
  outs.metadata,
  multiome,
  spatial,
  spatial_cluster = "RNACluster",
  multiome_cluster = "RNA_Clusters_HC",
  width = 4,
  height = 3,
  plot_dir = NULL
)
}
\arguments{
\item{outs.metadata}{A output from `run_projection()`, containing spatial coordinate of multiome cells}

\item{multiome}{A multiome Seurat object}

\item{spatial}{A spatial Seurat object}

\item{spatial_cluster}{column name in meta.data of spatial Seurat object containing cluster annotation of spatial data}

\item{multiome_cluster}{column name in meta.data of multiome Seurat object containing cluster annotation of multiome data}

\item{width}{width of the plot to save}

\item{height}{height of the plot to save}

\item{plot_dir}{Directory to save plots. If NULL, print plots. Example: '/directory' (instead of '/directory/')}
}
\value{
A heatmap showing the absolute and relative overlap of cells between two spatial and multiome clusters
}
\description{
Visualize the proportion of cells in multiome and spatial clusters.
Visualize the proportion of cells in multiome clusters that lie in spatial clusters, and vice versa.
}
