% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatialAutocorrelation.R
\name{ComputeAutocorrelation}
\alias{ComputeAutocorrelation}
\title{Compute autocorrelation}
\usage{
ComputeAutocorrelation(data, coords)
}
\arguments{
\item{data}{gene x cell matrix containing expression values}

\item{coords}{cell x 2 matrix containing spatial coordinate of cells}
}
\value{
A gene x gene matrix showing spatial autocorrelation represented by
global bivariate Moran's index
}
\description{
Compute spatial autocorrelation of genes
}
