% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCoexpression.R
\name{ComputeCoexpression}
\alias{ComputeCoexpression}
\title{ComputeCoexpression}
\usage{
ComputeCoexpression(data)
}
\arguments{
\item{data}{gene x cell matrix containing expression values}
}
\value{
A gene x gene matrix showing Pearson correlation coefficient
}
\description{
Compute correlation across genes of interest
}
