% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatiallyVariableFeatures.R
\name{ComputeRMetric}
\alias{ComputeRMetric}
\title{ComputeRMetric}
\usage{
ComputeRMetric(mv, r.metric = 5)
}
\arguments{
\item{mv}{Results of running markvario}

\item{r.metric}{r value at which to report the "trans" value of the markvariogram}
}
\value{
Returns a data.frame with r.metric values
}
\description{
Computes the metric at a given r (radius) value and stores
in meta.features from Seurat preprocessing functions
}
