% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatiallyVariableFeatures.R
\name{ComputeSpatiallyVariableFeatures}
\alias{ComputeSpatiallyVariableFeatures}
\title{Compute SpatiallyV ariable Features}
\usage{
ComputeSpatiallyVariableFeatures(data, coords, r.metric = 5)
}
\arguments{
\item{data}{gene x cell matrix containing expression values}

\item{coords}{cell x 2 matrix containing spatial coordinate of cells}

\item{r.metric}{r value at which to report the "trans" value of the markvariogram}
}
\value{
A dataframe showing mark variogram values, Moran's I values, and
gene ranks based on their spatial variability.
}
\description{
Compute spatially variable features from assay by
mark variogram and Moran's I computation
}
