% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunProjection.R
\name{MergeDatasets}
\alias{MergeDatasets}
\title{MergeDatasets}
\usage{
MergeDatasets(
  spatial,
  multiome,
  spatial_assay = "Xenium",
  multiome_assay = "RNA",
  spatial_cluster = "RNACluster",
  multiome_cluster = "RNA_Label",
  coord,
  harmony_lambda = NULL,
  gene_counts = 0,
  cell_counts = 0,
  plot_dir = NULL
)
}
\arguments{
\item{spatial}{A spatial Seurat object}

\item{multiome}{A multiome Seurat object}

\item{spatial_assay}{assay within spatial object shared with multiome data to integrate}

\item{multiome_assay}{assay within multiome object shared with spatial data to integrate}

\item{spatial_cluster}{column name in meta.data of spatial Seurat object containing cluster annotation of spatial data}

\item{multiome_cluster}{column name in meta.data of multiome Seurat object containing cluster annotation of multiome data}

\item{coord}{cell x 2 matrix containing spatial coordinate of cells}

\item{harmony_lambda}{Ridge regression penalty parameter.Default lambda=NULL for automatic lambda estimation.
Lambda must be strictly positive. Smaller values result in more aggressive correction.}

\item{gene_counts}{A threshold to remove genes based on summed gene counts. Default: 0.
i.e., only retain genes having summed gene counts > gene_counts.}

\item{cell_counts}{A threshold to remove cells based on summed cell counts. Default: 0.
i.e., only retain cells having summed cell counts > cell_counts}

\item{plot_dir}{Directory to save plots. If NULL, print plots.}
}
\value{
A list of integrated Seurat object from spatial and multiome data
and coordinate after filtering for cells with low counts
}
\description{
Integrate spatial and multiome datasets
}
