% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertintoSpatialObject.R
\name{MultiometoSpatialObject}
\alias{MultiometoSpatialObject}
\title{MultiometoSpatialObject}
\usage{
MultiometoSpatialObject(
  outs.metadata,
  spatial,
  multiome,
  spatial_assay = "MERSCOPE",
  image = "AAH_COLON_6_MONTHS"
)
}
\arguments{
\item{outs.metadata}{A output from `run_projection()`, containing spatial coordinate of multiome cells}

\item{spatial}{A spatial Seurat object}

\item{multiome}{A multiome Seurat object}

\item{spatial_assay}{assay within spatial object shared with multiome data to integrate}

\item{image}{am image (fov) from spatial data to include in the multiome spatial object}
}
\value{
A spatial Seurat object
}
\description{
Processing function to convert a multiome Seurat object into a spatial Seurat object by
adding spatial coordinates and image object
}
