% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatialNiche.R
\name{PlotCellTypeEnrichmentBarPlot}
\alias{PlotCellTypeEnrichmentBarPlot}
\title{PlotCellTypeEnrichmentBarPlot}
\usage{
PlotCellTypeEnrichmentBarPlot(niche.outs, CellType = 0)
}
\arguments{
\item{niche.outs}{- output table from `CellTypeNicheEnrichment`}

\item{CellType}{- which cell type to plot, must be present in the results}
}
\value{
barplot with enrichment results
}
\description{
Given niche results from CellTypeNicheEnrichment, will plot a niche enrichment barplot for any query cell type
}
