% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunProjection.R
\name{PlotCorrelation}
\alias{PlotCorrelation}
\title{PlotCorrelation}
\usage{
PlotCorrelation(
  spatial,
  multiome,
  spatial_assay = "Xenium",
  multiome_assay = "RNA",
  log10 = TRUE,
  plot_dir = NULL
)
}
\arguments{
\item{spatial}{A spatial Seurat object}

\item{multiome}{A multiome Seurat object}

\item{spatial_assay}{assay within spatial object shared with multiome data to integrate}

\item{multiome_assay}{assay within multiome object shared with spatial data to integrate}

\item{log10}{whether to plot correlation on after log10 transformation, default = TRUE}

\item{plot_dir}{Directory to save plots. If NULL, print plots.}
}
\value{
A scatterplot showing summed gene counts from spatial and multome data
}
\description{
A plotting function showing correlation of gene expression from spatial and multiome data
}
