% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePerformance.R
\name{PlotCorrelationSpatial}
\alias{PlotCorrelationSpatial}
\title{Plot correlation on spatial coordinate}
\usage{
PlotCorrelationSpatial(
  outs.metadata,
  multiome,
  spatial,
  spatial_assay = "Xenium",
  multiome_assay = "RNA",
  plot_dir = NULL
)
}
\arguments{
\item{outs.metadata}{A output from `run_projection()`, containing spatial coordinate of multiome cells}

\item{multiome}{A multiome Seurat object}

\item{spatial}{A spatial Seurat object}

\item{spatial_assay}{assay within spatial object shared with multiome data to integrate}

\item{multiome_assay}{assay within multiome object shared with spatial data to integrate}

\item{plot_dir}{Directory to save plots. If NULL, print plots. Example: '/directory' (instead of '/directory/')}
}
\value{
A dataframe showing Spearman and Pearson correlation of each multiome cells
}
\description{
Plot correlation in gene expression between spatial-multiome cell counterparts
on their spatial coordinates
}
