% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePerformance.R
\name{PlotGeneCorrelation}
\alias{PlotGeneCorrelation}
\title{Plot Gene Correlation}
\usage{
PlotGeneCorrelation(
  outs.metadata,
  gene = "CD3D",
  multiome,
  spatial,
  spatial_assay = "Xenium",
  multiome_assay = "RNA",
  pt.size = 0.3,
  plot_dir = NULL
)
}
\arguments{
\item{outs.metadata}{A output from `run_projection()`, containing spatial coordinate of multiome cells}

\item{gene}{gene of interest}

\item{multiome}{A multiome Seurat object}

\item{spatial}{A spatial Seurat object}

\item{spatial_assay}{assay within spatial object shared with multiome data to integrate}

\item{multiome_assay}{assay within multiome object shared with spatial data to integrate}

\item{pt.size}{size of points in scatter plot}

\item{plot_dir}{Directory to save plots. If NULL, print plots. Example: '/directory' (instead of '/directory/')}
}
\value{
A scatterplot showing gene expression in spatial and multiome data
}
\description{
Plot correlation in gene expression between spatial-multiome cell counterparts.
}
