% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatiallyVariableFeatures.R
\name{PlotMoransiAcrossModality}
\alias{PlotMoransiAcrossModality}
\title{Plot Moransi Across Modality}
\usage{
PlotMoransiAcrossModality(
  outs,
  plot.moransI = NA,
  top.features = 100,
  text.max.overlap = 10
)
}
\arguments{
\item{outs}{output dataframe from `compute_spatially_variable_features()`}

\item{plot.moransI}{A moran's I value threshold to plot genes e.g. 0.3.
Higher the moran's I value, more spatially variable. Default: NA}

\item{top.features}{A number of top spatially variable genes to plot. Default:100
If both plot.moransI = NA & top.features = NA, plot all genes.}

\item{text.max.overlap}{Exclude text labels when they overlap too many other things.
For each text label, we count how many other text labels or other data points it overlaps,
and exclude the text label if it has too many overlaps. Defaults to 10.}
}
\value{
A scatterplot showing moran's I value from assay 1 in x-axis and
assay 2 in y-axis
}
\description{
Plot moran's I values from two assays
}
