% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatialNiche.R
\name{PlotNicheEnrichmentBarPlot}
\alias{PlotNicheEnrichmentBarPlot}
\title{PlotNicheEnrichmentBarPlot}
\usage{
PlotNicheEnrichmentBarPlot(niche.outs, niche = 0)
}
\arguments{
\item{niche.outs}{- output table from `CellTypeNicheEnrichment`}

\item{niche}{- which niche to plot, must be present in the results}
}
\value{
barplot with enrichment results

 @export
}
\description{
given niche results from CellTypeNicheEnrichment, will plot a cell type enrichment barplot for any query niche
}
