% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatialNiche.R
\name{PlotNicheEnrichmentHeatmap}
\alias{PlotNicheEnrichmentHeatmap}
\title{PlotNicheEnrichmentHeatmap}
\usage{
PlotNicheEnrichmentHeatmap(niche.outs)
}
\arguments{
\item{niche.outs}{- output table from `CellTypeNicheEnrichment` function}
}
\value{
heatmap with enrichment results
}
\description{
Given niche results from cellTypeNicheEnrichment, will plot an overview heatmap of cell type to cell niche enrichment.
}
