% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePerformance.R
\name{PlotNNDistance}
\alias{PlotNNDistance}
\title{Plot nearest neighbor distance}
\usage{
PlotNNDistance(outs.metadata, plot_dir = NULL)
}
\arguments{
\item{outs.metadata}{A output from `run_projection()`, containing spatial coordinate of multiome cells}

\item{plot_dir}{Directory to save plots. If NULL, print plots. Example: '/directory' (instead of '/directory/')}
}
\value{
Diagnostic plots describing how close spatial counterparts were to each multiome cells
}
\description{
Diagnostic plot showing distance of closest spatial counterpart with multiome cells
}
