% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePerformance.R
\name{PlotSpatialAssignment}
\alias{PlotSpatialAssignment}
\title{Plot spatial assignment}
\usage{
PlotSpatialAssignment(outs.metadata, plot_dir = NULL)
}
\arguments{
\item{outs.metadata}{A output from `run_projection()`, containing spatial coordinate of multiome cells}

\item{plot_dir}{Directory to save plots. If NULL, print plots. Example: '/directory' (instead of '/directory/')}
}
\value{
A dataframe showing number of multiome cells assigned for unique spatial coordinate.
Plots histogram and scatterplot describing the number of multiome cells assigned to each spatial coordinate
}
\description{
Diagnostic plot showing number of multiome cells assigned to each spatial coordinate
}
