% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatiallyVariableFeatures.R
\name{PlotSpatialRankAcrossModality}
\alias{PlotSpatialRankAcrossModality}
\title{Plot Spatial Rank Across Modality}
\usage{
PlotSpatialRankAcrossModality(outs, plot.rank = NA, top.features = 100)
}
\arguments{
\item{outs}{output dataframe from `compute_spatially_variable_features()`}

\item{plot.rank}{A rank threshold to plot genes e.g. 100.
Lower the rank, more spatially variable. Default: NA}

\item{top.features}{A number of top spatially variable genes to plot. Default:100
If both plot.rank = NA & top.features = NA, plot all genes.}
}
\value{
A scatterplot showing ranks from assay 1 in x-axis and
assay 2 in y-axis
}
\description{
Plot spatially variable rank from two assays
}
