% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeSpatiallyVariableFeatures.R
\name{PlotSpatialVariableRanks}
\alias{PlotSpatialVariableRanks}
\title{PlotSpatialVariableRanks}
\usage{
PlotSpatialVariableRanks(outs, rank.threshold = 100, text.max.overlap = 10)
}
\arguments{
\item{outs}{output dataframe from `compute_spatially_variable_features()`}

\item{rank.threshold}{A rank threshold to plot genes. Default:100
i.e. showing genes having ranks equal or less than rank.threshold.
rank.threshold = NA plots all genes}

\item{text.max.overlap}{Exclude text labels when they overlap too many other things.
For each text label, we count how many other text labels or other data points it overlaps,
and exclude the text label if it has too many overlaps. Defaults to 10.}
}
\value{
A scatterplot showing mark variogram rank on x-axis and
moran's I rank on y-axis
}
\description{
Plot rank of each gene by their mark variogram and moran's I ranks.
}
