% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluatePerformance.R
\name{PlotSummedExpressionSpatial}
\alias{PlotSummedExpressionSpatial}
\title{PlotSummedExpressionSpatial}
\usage{
PlotSummedExpressionSpatial(
  outs.metadata,
  multiome,
  spatial,
  spatial_assay = "Xenium",
  multiome_assay = "RNA",
  plot_dir = NULL
)
}
\arguments{
\item{outs.metadata}{A output from `run_projection()`, containing spatial coordinate of multiome cells}

\item{multiome}{A multiome Seurat object}

\item{spatial}{A spatial Seurat object}

\item{spatial_assay}{assay within spatial object shared with multiome data to integrate}

\item{multiome_assay}{assay within multiome object shared with spatial data to integrate}

\item{plot_dir}{Directory to save plots. If NULL, print plots. Example: '/directory' (instead of '/directory/')}
}
\value{
A dataframe showing summed gene counts of each multiome cells
}
\description{
Plot summed gene counts for each multiome cells on their spatial coordinates
}
