% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunProjection.R
\name{RunProjection}
\alias{RunProjection}
\title{RunProjection}
\usage{
RunProjection(
  merge,
  multiome,
  coord_dt,
  knn_k = 100,
  n_harmony_dim = 35,
  multiome_cluster = "RNA_Label",
  plot_dir = NULL
)
}
\arguments{
\item{merge}{A merged Seurat object, output from `MergeDatasets()` function, merging spatial and multiome datasets}

\item{multiome}{A multiome Seurat object}

\item{coord_dt}{cell x 3 data frame containing cell names and spatial coordinate of cells, output from `MergeDatasets()`}

\item{knn_k}{a cutoff for the top k k-nearest neighbors to find a spatial counterpart for. Default: 100.
i.e., if there is no spatial counterpart in the top 100 neighbors, no spatial coordinate
is assigned to the respective multiome cell.
coordinate is assigned for the respective multiome cell.}

\item{n_harmony_dim}{Number of harmony dimension to use to compute nearest neighbour. Default: 35}

\item{multiome_cluster}{column name in meta.data of merge Seurat object to visualise after projection}

\item{plot_dir}{Directory to save plots. If NULL, print plots.}
}
\value{
a metadata containing spatial coordinate of multliome cells
}
\description{
Core function to integrate spatial and multiome data, and assign spatial coordinate to multiome cells
}
