% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spamsc.R
\name{spamsc}
\alias{spamsc}
\title{spamsc: A package for a comprehensive workflow designed to project multi-modal
single-cell data onto high-resolution spatial transcriptomic data and identify spatial features.}
\description{
The spamsc package contains functions to
}
\details{
It addresses the following needs:
\itemize{ \item Check if multi-modal single-cell and spatial data can be integrated;
\item Spatially project multi-modal data;
\item Diagnose and evaluate projection;
\item Identify spatial features;
}
}
