% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCoexpression.R
\name{VisualiseCoexpression}
\alias{VisualiseCoexpression}
\title{VisualiseCoexpression}
\usage{
VisualiseCoexpression(cor_mtx, cor.threshold = 0.4)
}
\arguments{
\item{cor_mtx}{output correlation matrix from `compute_coexpression()`}

\item{cor.threshold}{A minimum Pearson correlation coefficient threshold to plot genes. Default: 0.4}
}
\value{
A heatmap showing Perason correalaiton coefficient across genes
}
\description{
Plot heatmap showing the correlation between genes.
}
