% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComputeCoexpression.R
\name{VisualiseCoexpressionWindow}
\alias{VisualiseCoexpressionWindow}
\title{Visualise coexpression within range}
\usage{
VisualiseCoexpressionWindow(
  cor_mtx,
  cor.top.threshold = 0.8,
  cor.bottom.threshold = 0.4
)
}
\arguments{
\item{cor_mtx}{output correlation matrix from `compute_coexpression()`}

\item{cor.top.threshold}{A maximum Pearson correlation coefficient threshold to plot genes. Default: 0.4}

\item{cor.bottom.threshold}{A minimum Pearson correlation coefficient threshold to plot genes. Default: 0.4}
}
\value{
A heatmap showing Perason correalaiton coefficient across genes
}
\description{
Plot heatmap showing the correlation between genes.
}
