% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualiseProjection.R
\name{VisualiseFeatures}
\alias{VisualiseFeatures}
\title{VisualiseFeatures}
\usage{
VisualiseFeatures(
  outs.metadata,
  multiome,
  feature = NA,
  pt.size = 0.4,
  order = FALSE,
  plot_dir = NULL
)
}
\arguments{
\item{outs.metadata}{A output from `run_projection()`, containing spatial coordinate of multiome cells}

\item{multiome}{A multiome Seurat object}

\item{feature}{feature of interest e.g. RNA counts, cluster labels}

\item{pt.size}{size of points on plot. Default = 0.5}

\item{order}{order cells by the gene expression}

\item{plot_dir}{Directory to save plots. If NULL, print plots. Example: '/directory' (instead of '/directory/')}
}
\value{
A scatterplot showing feature on spatial coordinates.
}
\description{
Visualize feature expression from meta.data on spatial coordinates
}
