% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualiseProjection.R
\name{VisualisePairedModalityOneCounterpart}
\alias{VisualisePairedModalityOneCounterpart}
\title{Visualise Paired Modality One Counterpart}
\usage{
VisualisePairedModalityOneCounterpart(
  outs.metadata,
  lookup_table,
  multiome,
  spatial,
  gene,
  multiome_assay = "RNA",
  multiome_project_assay = "ADT",
  pt.size = 0.5,
  plot_dir = NULL
)
}
\arguments{
\item{outs.metadata}{A output from `run_projection()`, containing spatial coordinate of multiome cells}

\item{lookup_table}{A table containing features of the multiome assay used for integration (named 'gene')
and the corresponding features in the assay to project (named 'paired'). If the feature names are the same,
duplicate the column with the names 'gene' and 'paired'.}

\item{multiome}{A multiome Seurat object}

\item{spatial}{A spatial Seurat object}

\item{gene}{gene of interest}

\item{multiome_assay}{assay within multiome object shared with spatial data}

\item{multiome_project_assay}{assay within multiome object to project}

\item{pt.size}{size of points on plot. Default = 0.5}

\item{plot_dir}{Directory to save plots. If NULL, print plots. Example: '/directory' (instead of '/directory/')}
}
\value{
A scatterplot showing gene expression in the shared and projected assays on spatial coordinates.
}
\description{
Visualize gene expression from twp multiome assays on spatial coordinates
when there is only one counterpart between assay 1 (e.g., RNA assay) and assay 2 (e.g., protein assay).
}
