% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VisualiseProjection.R
\name{VisualiseSingleModality}
\alias{VisualiseSingleModality}
\title{VisualiseSingleModality}
\usage{
VisualiseSingleModality(
  outs.metadata,
  multiome,
  gene = NA,
  multiome_plot_assay = "GeneScore",
  pt.size = 0.4,
  order = FALSE,
  plot_dir = NULL,
  plot_height = 3,
  plot_width = 4
)
}
\arguments{
\item{outs.metadata}{A output from `run_projection()`, containing spatial coordinate of multiome cells}

\item{multiome}{A multiome Seurat object}

\item{gene}{gene of interest}

\item{multiome_plot_assay}{assay within multiome object to plot}

\item{pt.size}{size of points on plot. Default = 0.5}

\item{order}{order cells by the gene expression}

\item{plot_dir}{Directory to save plots. If NULL, print plots. Example: '/directory' (instead of '/directory/')}
}
\value{
A scatterplot showing gene expression from any mulitome assay on spatial coordinates.
}
\description{
Visualize gene expression from any multiome assays on spatial coordinates
}
