% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunProjection.R
\name{RunProjection}
\alias{RunProjection}
\title{RunProjection}
\usage{
RunProjection(
  merge,
  coord,
  lambda = 0.4,
  knn_k = 100,
  n_harmony_dim = 35,
  plot_dir = NULL
)
}
\arguments{
\item{merge}{A merged Seurat object, output from `MergeDatasets()` function, merging spatial and multiome datasets}

\item{coord}{cell x 2 matrix containing spatial coordinate of cells}

\item{lambda}{Ridge regression penalty parameter for Harmony integration of spatial and multiome datasets.
Default lambda=0.4. Lambda must be strictly positive. Smaller values result in more aggressive correction.}

\item{knn_k}{a cutoff for the top k k-nearest neighbors to find a spatial counterpart for. Default: 100.
i.e., if there is no spatial counterpart in the top 100 neighbors, no spatial coordinate
is assigned to the respective multiome cell.
coordinate is assigned for the respective multiome cell.}

\item{n_harmony_dim}{Number of harmony dimension to use to compute nearest neighbour. Default: 35}

\item{plot_dir}{Directory to save plots. If NULL, print plots.}
}
\value{
a metadata containing spatial coordinate of multliome cells
}
\description{
Core function to integrate spatial and multiome data, and assign spatial coordinate to multiome cells
}
