% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunProjection.R
\name{RunProjection}
\alias{RunProjection}
\title{Run projection}
\usage{
RunProjection(
  spatial,
  multiome,
  spatial_assay = "Xenium",
  multiome_assay = "RNA",
  spatial_cluster = "RNACluster",
  multiome_cluster = "RNA_Label",
  coord,
  lambda = 0.4,
  knn_k = 100,
  n_harmony_dim = 35,
  gene_counts = 0,
  cell_counts = 0,
  plot_dir = NULL
)
}
\arguments{
\item{spatial}{A spatial Seurat object}

\item{multiome}{A multiome Seurat object}

\item{spatial_assay}{assay within spatial object shared with multiome data to integrate}

\item{multiome_assay}{assay within multiome object shared with spatial data to integrate}

\item{spatial_cluster}{column name in meta.data of spatial Seurat object containing cluster annotation of spatial data}

\item{multiome_cluster}{column name in meta.data of multiome Seurat object containing cluster annotation of multiome data}

\item{coord}{cell x 2 matrix containing spatial coordinate of cells}

\item{lambda}{Ridge regression penalty parameter for Harmony integration of spatial and multiome datasets.
Default lambda=0.4. Lambda must be strictly positive. Smaller values result in more aggressive correction.}

\item{knn_k}{a cutoff for the top k k-nearest neighbors to find a spatial counterpart for. Default: 100.
i.e., if there is no spatial counterpart in the top 100 neighbors, no spatial coordinate
is assigned to the respective multiome cell.
coordinate is assigned for the respective multiome cell.}

\item{n_harmony_dim}{Number of harmony dimension to use to compute nearest neighbour. Default: 35}

\item{gene_counts}{A threshold to remove genes based on summed gene counts. Default: 0.
i.e., only retain genes having summed gene counts > gene_counts.}

\item{cell_counts}{A threshold to remove cells based on summed cell counts. Default: 0.
i.e., only retain cells having summed cell counts > cell_counts}

\item{plot_dir}{Directory to save plots. If NULL, print plots. Example: '/directory' (instead of '/directory/')}
}
\value{
A list of objects: a metadata containing spatial coordinate of multliome cells and
an integrated seurat object from spatial and multiome data
}
\description{
Core function to integrate spatial and multiome data, and assign spatial coordinate to multiome cells
}
